refactor: move platform path logic
